pyproject.toml
[tool.poetry]
name = "simple-chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.27"
psycopg2 = "^2.9.9"
passlib = "^1.7.4"
email-validator = "^2.1.1"
alembic = "^1.13.1"
langchain = "^0.1.9"
langchain-openai = "^0.0.8"
psycopg = "^3.1.18"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
